requirements.txt
numpy==1.17.1
pandas
streamlit
plotly.express
